+++ changelog.md
@@ -18,6 +18,7 @@
 * added [ServiceRedirectionAction](https://github.com/Unicon/cas-addons/blob/master/src/main/java/net/unicon/cas/addons/web/flow/ServiceRedirectionAction.java)
 * added [InMemoryServiceRedirectionByClientIpAddressAdvisor](https://github.com/Unicon/cas-addons/blob/master/src/main/java/net/unicon/cas/addons/web/flow/InMemoryServiceRedirectionByClientIpAddressAdvisor.java)
 * added custom XML schema element (`service-redirection-action`) `BeanDefinitionParser` implementation for `ServiceRedirectionAction`
+* added custom XML schema element (`request-param-login-view-selector`) `BeanDefinitionParser` implementation for `RequestPararameterCasLoginViewSelector`
 
 
 ###Changes in version 1.8 (2013-09-09)